+++ src/Networking/motor_interface.cpp
@@ -64,7 +64,7 @@ bool motorSupportsPID(int motor_serial)
 
 int getIndex(const std::vector<std::string> &arr, const std::string &value)
 {
-  for (int i = 0; i < arr.size(); i++) {
+  for (size_t i = 0; i < arr.size(); i++) {
     if (arr[i] == value) {
       return i;
     }
