+++ src/worldmap/TrICP.cpp
@@ -145,7 +145,7 @@ bool TrICP::isDone(int numIter, double mse, double oldMSE) const
 transform_t TrICP::iterate(points_t &sample, int N, double &mse) const
 {
 	PointPair pairs[sample.size()];
-	for (int i = 0; i < sample.size(); i++)
+	for (size_t i = 0; i < sample.size(); i++)
 	{
 		const point_t &point = sample[i];
 		point_t closestPoint = getClosest(point);
